requirements.txt
python-dotenv
openrouter
requests
groq
openai
